emacs: Update emacs-overlay and apply patch
Emacs commit ea9831bb3cb4878273f6f848051c9b8c3c76d5f1 introduces a
failure to build nix elisp packages on certain machines. Laptop has no
issue, but desktop fails to compile some packages. This is reproducible
inside a container so unlikely to be system state. Potentially could be
a but with libggcjit, as it does not occur with native comp disabled.

For now, this adds a patch that reverts part of the commit that triggers
the bug.